+++ src/interfaces/people/index.ts
@@ -1,9 +1,14 @@
+import { GenreInterface } from '../audible'
+
 interface Person {
     name: string;
 }
 
 export interface AuthorInterface extends Person {
-    asin?: string
+    asin?: string;
+    description?: string;
+    genres?: GenreInterface[];
+    image?: string;
 }
 
 export interface NarratorInterface extends Person {}