+++ tests/audible/books/audibleApi.test.ts
@@ -1,6 +1,6 @@
-import ApiHelper from '../../src/helpers/audibleApi'
-import { AudibleInterface } from '../../src/interfaces/audible'
-import { ApiBookInterface } from '../../src/interfaces/books'
+import ApiHelper from '../../../src/helpers/books/audibleApi'
+import { AudibleInterface } from '../../../src/interfaces/audible'
+import { ApiBookInterface } from '../../../src/interfaces/books'
 
 let asinBad: string
 let apiBad: ApiHelper